requirements.txt
altgraph==0.16.1
colorama==0.4.1
future==0.17.1
macholib==1.11
pefile==2018.8.8
Pillow==5.4.1
PyInstaller==3.4
pyperclip==1.7.0
pywin32-ctypes==0.2.0
qrcode==6.1
six==1.12.0
zxing==0.9.3
